pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "yoclip"
version = "0.1.0"
description = "A CLI tool built with Typer"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    { name = "Your Name", email = "your.email@example.com" },
]
keywords = ["cli", "typer", "command-line"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "torch",
    "torchvision",
    "transformers",
    "open-clip-torch",
    "typer",
    "rich",
    "Pillow",
    "pandas",
    "numpy",
    "scikit-learn",
    "rasterio",
    "geopandas",
    "shapely",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
yoclip = "yoclip.main:app"

[project.urls]
Homepage = "https://github.com/yourusername/yoclip"
Repository = "https://github.com/yourusername/yoclip"
Documentation = "https://github.com/yourusername/yoclip/blob/main/README.md"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=yoclip --cov-report=term-missing"
